+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `get-validator` and `set-validator!`
 - Fix a `def` or `deftype` inside a function being added to the module exports, which made the module fail to load
 - Fix `take-nth` to throw on a non-number step instead of looping forever
+- Fix `empty` to return nil on non-collections and class instances instead of throwing, like CLJS
 
 ## 0.14.201